core: collapse the shim methods that no longer vary by version
Six SparkShims methods have byte-identical overrides in all four remaining
shims. Moved the implementation into the trait and dropped the overrides.

createParquetFilters stays in the shims: LegacyBehaviorPolicy sits inside
SQLConf on 3.4 and at top level from 3.5 on, so no single import in
shims/common satisfies all four versions.
